+++ src/styles/App.css
@@ -114,7 +114,7 @@
   border-radius: 10px;
   height: fit-content;
   width: fit-content;
-  zoom: 110%;
+
   justify-content: center;
 }
